force 32-bit for now with -m32. Added nifty debug info each time a control msg is sent or received. Fixed the IROB assignment (off-by-one).
git-svn-id: file:///home/svnshare/juggler_2.6.22.14/libcmm/trunk@406 1971e5c1-a347-0410-b49b-82492f0cf60c

+++ libcmm_test_sender.cpp
@@ -62,7 +62,7 @@ int get_reply(mc_socket_t sock)
         return rc;
     }
     ch.data[sizeof(ch)-1] = '\0';
-    fprintf(stderr, "Echo: %*s\n", sizeof(ch) - 1, ch.data);
+    fprintf(stderr, "Echo: %*s\n", (int)(sizeof(ch) - 1), ch.data);
     return rc;
 }
